Use parameterized tests for network testing
