+++ gradle/versions.gradle
@@ -15,10 +15,10 @@ ext {
     coreLibraryDesugaringVersion = "1.0.9"
 
     // Libraries
-    androidxTestRunnerVersion = "1.2.0"
-    androidxTestOrchestratorVersion = "1.2.0"
-    androidxTestMonitorVersion = "1.2.0"
-    androidxTestCoreVersion = "1.2.0"
+    androidxTestRunnerVersion = "1.3.0"
+    androidxTestOrchestratorVersion = "1.3.0"
+    androidxTestMonitorVersion = "1.3.0"
+    androidxTestCoreVersion = "1.3.0"
     androidxJunitVersion = "1.1.1"
     annotationVersion = "1.0.0"
     appCompatVersion = "1.0.2"
