+++ msalautomationapp/src/androidTest/resources/com/microsoft/identity/client/msal/automationapp/testpass/network/input_acquireTokenWithoutBroker.csv
@@ -13,38 +13,3 @@ TEST12,WIFI,4,WIFI,6,CELLULAR,4,WIFI_AND_CELLULAR,7,CELLULAR,8,CELLULAR,8,WIFI,7
 TEST13,NONE,0,WIFI_AND_CELLULAR,9,WIFI,4,WIFI,0,WIFI_AND_CELLULAR,3,WIFI_AND_CELLULAR,4,WIFI_AND_CELLULAR,0,WIFI_AND_CELLULAR,7,WIFI_AND_CELLULAR,8,WIFI,6,NONE,1,CELLULAR,4,CELLULAR,7,WIFI,7
 TEST14,CELLULAR,0,WIFI,4,WIFI_AND_CELLULAR,2,CELLULAR,2,NONE,6,WIFI,4,WIFI_AND_CELLULAR,9,NONE,3,WIFI,2,WIFI_AND_CELLULAR,5,WIFI_AND_CELLULAR,0,WIFI_AND_CELLULAR,9,NONE,5,CELLULAR,0,WIFI,6,WIFI,3
 TEST15,NONE,5,CELLULAR,2,WIFI,4,WIFI_AND_CELLULAR,2,CELLULAR,6,NONE,1,WIFI,4,WIFI,0,WIFI,8,NONE,9,WIFI,1,WIFI_AND_CELLULAR,0,WIFI_AND_CELLULAR,5,WIFI,9,WIFI_AND_CELLULAR,4
-TEST16,WIFI,4,CELLULAR,2,CELLULAR,5,CELLULAR,8,CELLULAR,2,WIFI_AND_CELLULAR,9,NONE,1,WIFI,0,CELLULAR,2,NONE,5,CELLULAR,8,CELLULAR,5,CELLULAR,8,NONE,1
-TEST17,CELLULAR,5,WIFI,4,WIFI_AND_CELLULAR,8,CELLULAR,4,WIFI,9,NONE,9,CELLULAR,9,NONE,3,WIFI,4,WIFI,0,CELLULAR,5
-TEST18,WIFI,5,WIFI_AND_CELLULAR,1,WIFI,2,WIFI_AND_CELLULAR,0,WIFI_AND_CELLULAR,9,WIFI,2,CELLULAR,0,NONE,1,CELLULAR,4,WIFI,9,CELLULAR,9,NONE,5,WIFI_AND_CELLULAR,1,WIFI,8,WIFI,4
-TEST19,WIFI_AND_CELLULAR,1,WIFI,3,WIFI_AND_CELLULAR,1,WIFI,2,CELLULAR,4,CELLULAR,9,CELLULAR,7,WIFI_AND_CELLULAR,2,WIFI,1,NONE,7,WIFI,0,WIFI,4,CELLULAR,2,WIFI_AND_CELLULAR,9,WIFI_AND_CELLULAR,2,WIFI_AND_CELLULAR,3,NONE,1,CELLULAR,2
-TEST20,WIFI,1,NONE,6,WIFI,5,WIFI_AND_CELLULAR,3,NONE,9,CELLULAR,9,WIFI,3,CELLULAR,6,WIFI_AND_CELLULAR,9,CELLULAR,2,WIFI,1,CELLULAR,2,WIFI_AND_CELLULAR,4
-TEST21,CELLULAR,9,WIFI,2,WIFI,2,NONE,3,NONE,0,CELLULAR,0,WIFI_AND_CELLULAR,0,WIFI_AND_CELLULAR,6,WIFI,5,WIFI_AND_CELLULAR,0,CELLULAR,2,CELLULAR,1,NONE,9,WIFI,7,WIFI_AND_CELLULAR,8,WIFI_AND_CELLULAR,6
-TEST22,WIFI_AND_CELLULAR,7,WIFI,4,CELLULAR,4,WIFI,7,CELLULAR,7,WIFI_AND_CELLULAR,3,WIFI,9,NONE,0,CELLULAR,7,NONE,7,CELLULAR,5
-TEST23,NONE,9,WIFI,1,WIFI,1,NONE,7,WIFI_AND_CELLULAR,5,WIFI_AND_CELLULAR,5,CELLULAR,5,WIFI,2,WIFI,8,CELLULAR,5,WIFI_AND_CELLULAR,1,WIFI,2,NONE,6,NONE,0,NONE,0,WIFI_AND_CELLULAR,1,CELLULAR,2
-TEST24,WIFI,2,WIFI,9,NONE,4,CELLULAR,3,CELLULAR,5,CELLULAR,2,WIFI,7,WIFI_AND_CELLULAR,7,CELLULAR,8,WIFI,6,WIFI_AND_CELLULAR,7
-TEST25,CELLULAR,4,WIFI_AND_CELLULAR,4,NONE,9,NONE,7,NONE,0,CELLULAR,2,WIFI_AND_CELLULAR,3,WIFI_AND_CELLULAR,7,WIFI,5,CELLULAR,5,WIFI_AND_CELLULAR,4,CELLULAR,4,WIFI_AND_CELLULAR,2,NONE,4
-TEST26,WIFI_AND_CELLULAR,8,WIFI,9,CELLULAR,2,NONE,7,WIFI_AND_CELLULAR,9,CELLULAR,3,NONE,8,WIFI,1,WIFI,6,CELLULAR,2,NONE,5
-TEST27,CELLULAR,3,WIFI,0,CELLULAR,0,CELLULAR,5,WIFI_AND_CELLULAR,0,CELLULAR,5,WIFI_AND_CELLULAR,2,CELLULAR,9,WIFI,1,CELLULAR,7,NONE,3,WIFI_AND_CELLULAR,5,WIFI_AND_CELLULAR,1,WIFI,5,WIFI,0,NONE,4,CELLULAR,9,NONE,1
-TEST28,CELLULAR,0,WIFI_AND_CELLULAR,4,WIFI_AND_CELLULAR,6,CELLULAR,1,CELLULAR,6,WIFI_AND_CELLULAR,2,NONE,0,NONE,6,WIFI,2,NONE,6,WIFI_AND_CELLULAR,2,NONE,8,WIFI,6,CELLULAR,8,WIFI,3
-TEST29,NONE,8,CELLULAR,9,WIFI_AND_CELLULAR,2,CELLULAR,3,CELLULAR,4,WIFI_AND_CELLULAR,5,CELLULAR,3,WIFI,0,WIFI,4,CELLULAR,2,CELLULAR,5,NONE,4,NONE,0,CELLULAR,8,WIFI,3
-TEST30,WIFI,5,WIFI,5,WIFI_AND_CELLULAR,6,WIFI,2,WIFI_AND_CELLULAR,1,WIFI_AND_CELLULAR,9,NONE,0,WIFI,0,WIFI,7,CELLULAR,3,NONE,8,NONE,9,CELLULAR,1,CELLULAR,3,CELLULAR,1
-TEST31,NONE,0,WIFI_AND_CELLULAR,7,NONE,8,WIFI_AND_CELLULAR,6,WIFI,8,WIFI,2,WIFI,3,WIFI,5,WIFI,2,CELLULAR,7,CELLULAR,4,WIFI,7,NONE,1
-TEST32,WIFI_AND_CELLULAR,2,WIFI_AND_CELLULAR,5,WIFI,5,WIFI,1,WIFI,1,WIFI_AND_CELLULAR,8,NONE,2,WIFI_AND_CELLULAR,0,CELLULAR,8,CELLULAR,2,WIFI_AND_CELLULAR,0,WIFI_AND_CELLULAR,7,WIFI_AND_CELLULAR,4,NONE,4,WIFI_AND_CELLULAR,8,WIFI_AND_CELLULAR,2,NONE,1
-TEST33,NONE,8,NONE,5,CELLULAR,0,WIFI_AND_CELLULAR,7,WIFI_AND_CELLULAR,1,WIFI_AND_CELLULAR,1,CELLULAR,5,CELLULAR,4,CELLULAR,6,NONE,0,NONE,0,CELLULAR,5,CELLULAR,3,NONE,4,WIFI,5,WIFI_AND_CELLULAR,5,NONE,1
-TEST34,CELLULAR,2,NONE,3,NONE,3,NONE,1,CELLULAR,7,NONE,1,CELLULAR,0,CELLULAR,4,WIFI_AND_CELLULAR,8,WIFI_AND_CELLULAR,2,NONE,5,WIFI_AND_CELLULAR,5,NONE,2,WIFI,3,WIFI_AND_CELLULAR,5,CELLULAR,7,NONE,2
-TEST35,WIFI,2,NONE,7,CELLULAR,4,WIFI_AND_CELLULAR,7,NONE,4,NONE,3,CELLULAR,3,NONE,1,WIFI_AND_CELLULAR,9,WIFI,4,WIFI,9,WIFI,7
-TEST36,WIFI_AND_CELLULAR,7,CELLULAR,5,WIFI_AND_CELLULAR,6,WIFI_AND_CELLULAR,8,WIFI,8,WIFI_AND_CELLULAR,4,WIFI_AND_CELLULAR,9,CELLULAR,7,WIFI,0,NONE,6
-TEST37,NONE,6,NONE,4,CELLULAR,1,WIFI_AND_CELLULAR,3,WIFI,0,WIFI_AND_CELLULAR,6,NONE,4,WIFI_AND_CELLULAR,6,WIFI,8,WIFI,6,NONE,9,WIFI_AND_CELLULAR,6,NONE,1
-TEST38,CELLULAR,0,WIFI_AND_CELLULAR,5,WIFI,7,NONE,5,WIFI,8,CELLULAR,6,WIFI,8,CELLULAR,6,WIFI,2,WIFI_AND_CELLULAR,7,WIFI_AND_CELLULAR,6
-TEST39,NONE,6,WIFI_AND_CELLULAR,1,CELLULAR,4,CELLULAR,8,CELLULAR,7,WIFI,0,WIFI,9,CELLULAR,6,NONE,4,NONE,8,CELLULAR,7
-TEST40,NONE,8,CELLULAR,8,CELLULAR,4,CELLULAR,4,NONE,2,WIFI,0,WIFI,7,WIFI_AND_CELLULAR,5,CELLULAR,1,NONE,8,NONE,0,WIFI,4,NONE,1,WIFI,2,WIFI_AND_CELLULAR,1,CELLULAR,0,NONE,5
-TEST41,NONE,3,WIFI,3,NONE,8,WIFI,3,NONE,0,WIFI,1,CELLULAR,8,CELLULAR,7,NONE,2,WIFI,3,NONE,6,WIFI,0,CELLULAR,4,WIFI_AND_CELLULAR,4,WIFI,8
-TEST42,WIFI,4,WIFI,5,NONE,8,CELLULAR,2,NONE,6,NONE,7,CELLULAR,3,CELLULAR,2,WIFI_AND_CELLULAR,9,WIFI_AND_CELLULAR,7,WIFI_AND_CELLULAR,2,NONE,5
-TEST43,WIFI_AND_CELLULAR,3,WIFI,8,NONE,5,CELLULAR,3,CELLULAR,1,WIFI_AND_CELLULAR,4,WIFI_AND_CELLULAR,0,WIFI_AND_CELLULAR,1,CELLULAR,5,CELLULAR,3,WIFI_AND_CELLULAR,5,NONE,6,NONE,2,NONE,3,NONE,6,WIFI,3,WIFI,2
-TEST44,WIFI_AND_CELLULAR,9,WIFI_AND_CELLULAR,5,CELLULAR,7,NONE,2,NONE,3,NONE,9,WIFI_AND_CELLULAR,8,WIFI_AND_CELLULAR,8,CELLULAR,3,NONE,4,WIFI,2
-TEST45,NONE,2,WIFI,9,WIFI_AND_CELLULAR,8,WIFI,2,WIFI,1,NONE,9,CELLULAR,6,WIFI,8,WIFI_AND_CELLULAR,1,WIFI_AND_CELLULAR,5,CELLULAR,4,WIFI,0,CELLULAR,3,CELLULAR,2
-TEST46,NONE,1,CELLULAR,0,WIFI_AND_CELLULAR,7,CELLULAR,4,WIFI,7,CELLULAR,1,NONE,8,NONE,9,CELLULAR,5,WIFI,3,WIFI,2,CELLULAR,8,WIFI_AND_CELLULAR,5
-TEST47,WIFI,6,CELLULAR,5,NONE,6,WIFI,3,WIFI_AND_CELLULAR,7,WIFI_AND_CELLULAR,7,CELLULAR,1,WIFI,8,WIFI_AND_CELLULAR,8,WIFI,4,WIFI_AND_CELLULAR,5
-TEST48,WIFI,0,WIFI,4,CELLULAR,9,WIFI,7,WIFI,3,WIFI,5,CELLULAR,9,WIFI_AND_CELLULAR,2,WIFI_AND_CELLULAR,1,NONE,3,WIFI,4,CELLULAR,5,CELLULAR,6,NONE,1,WIFI_AND_CELLULAR,1
-TEST49,NONE,0,NONE,4,CELLULAR,5,CELLULAR,4,WIFI_AND_CELLULAR,3,CELLULAR,0,NONE,4,CELLULAR,7,WIFI,8,WIFI_AND_CELLULAR,1,CELLULAR,1,WIFI,0,WIFI,2,WIFI_AND_CELLULAR,2,NONE,5,WIFI,8,WIFI,6
-TEST50,CELLULAR,8,CELLULAR,2,WIFI,7,CELLULAR,6,WIFI_AND_CELLULAR,9,WIFI,3,WIFI,4,WIFI,2,WIFI_AND_CELLULAR,9,NONE,9,WIFI,1
