+++ msalautomationapp/src/androidTest/resources/com/microsoft/identity/client/msal/automationapp/testpass/network/output_acquireTokenWithoutBroker.csv
@@ -1,50 +1,15 @@
-TEST1,Fail,com.microsoft.identity.client.exception.MsalClientException: An IO error occurred with message: Unable to resolve host "login.microsoftonline.com": No address associated with hostname,com.microsoft.identity.client.exception.MsalClientException
-TEST2,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST3,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
+TEST1,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$)
+TEST2,Fail,net::ERR_INTERNET_DISCONNECTED
+TEST3,Fail,Timed out waiting for: .*,
 TEST4,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
 TEST5,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
 TEST6,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
 TEST7,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST8,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
+TEST8,Fail,net::ERR_INTERNET_DISCONNECTED
 TEST9,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
 TEST10,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
 TEST11,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
 TEST12,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST13,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
+TEST13,Fail,net::ERR_INTERNET_DISCONNECTED
 TEST14,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST15,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST16,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST17,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST18,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST19,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST20,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST21,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST22,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST23,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST24,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST25,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST26,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST27,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST28,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST29,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST30,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST31,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST32,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST33,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST34,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST35,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST36,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST37,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST38,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST39,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST40,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST41,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST42,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST43,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST44,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST45,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST46,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST47,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST48,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST49,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
-TEST50,Pass,(^[\w-]*\.[\w-]*\.[\w-]*$),
+TEST15,Fail,net::ERR_INTERNET_DISCONNECTED
